feat: promote the crash handler into libcamp_crash (#217)
install_thread_alt_stack() has to be callable from the first statement of
every thread CAMP starts, and CAMP starts threads in camp_map and
camp_map_ros — which cannot call into the executable that links them.
Compiled into the executable only, those threads had no symbol to call and
no report for a stack-overflow SIGSEGV. Round 2 recorded that as a
follow-up; this closes it.

The split is mechanism from policy, and the reason is a dependency
boundary. camp_map is ROS-free by design (ADR-0002) and now links this
library, so anything libcamp_crash links, camp_map links. crash_log_path()
is the one piece of #217 that needs rclcpp, so it moves the other way —
into src/camp/crash_log_path.cpp, in the executable, which already depends
on rclcpp. Reimplementing rcl's ROS_LOG_DIR / ROS_HOME / ~/.ros precedence
to keep the library whole was considered and rejected: a second copy of
upstream policy, free to drift out of agreement with the directory
ros2 launch actually writes to, is a diagnostics feature failing quietly.

One namespace (camp_crash), two targets, for that stated reason.

test_crash_handler now links the shipped library rather than recompiling
its source: the artifact camp_map links is the one that has to work, and a
private recompile could stay green while the installed library did not.
